+++ src/DotBPE.Rpc/Internal/InternalHelper.cs
@@ -2,12 +2,12 @@ namespace DotBPE.Rpc.Internal
 {
     public static class InternalHelper
     {
-        public static string FormatServiceIdentity(string serviceGroupName, ushort serviceId, ushort messageId)
+        public static string FormatServiceIdentity(string serviceGroupName, int serviceId, ushort messageId)
         {
             return $"{serviceGroupName}.{serviceId}.{messageId}";
         }
 
-        public static string FormatServicePath(ushort serviceId, ushort messageId)
+        public static string FormatServicePath(int serviceId, ushort messageId)
         {
             return $"{serviceId}.{messageId}";
         }
